Move inner class (and also junit tests)
