--- src/main/java/io/github/jhipster/registry/web/rest/JWTToken.java
@@ -1,24 +1,0 @@
-package io.github.jhipster.registry.web.rest;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-/**
- * Object to return as body in JWT Authentication.
- */
-public class JWTToken {
-
-    private String idToken;
-
-    public JWTToken(String idToken) {
-        this.idToken = idToken;
-    }
-
-    @JsonProperty("id_token")
-    public String getIdToken() {
-        return idToken;
-    }
-
-    public void setIdToken(String idToken) {
-        this.idToken = idToken;
-    }
-}
